+++ components/src/jsMain/kotlin/dev/fritz2/components/modal.kt
@@ -1,8 +1,12 @@
 package dev.fritz2.components
 
-import dev.fritz2.binding.*
+import dev.fritz2.binding.RootStore
+import dev.fritz2.binding.SimpleHandler
+import dev.fritz2.binding.storeOf
+import dev.fritz2.binding.watch
 import dev.fritz2.dom.html.Div
 import dev.fritz2.dom.html.RenderContext
+import dev.fritz2.dom.html.Scope
 import dev.fritz2.styling.StyleClass
 import dev.fritz2.styling.div
 import dev.fritz2.styling.params.BasicParams
