+++ components/src/jsMain/kotlin/dev/fritz2/components/toast.kt
@@ -1,10 +1,10 @@
 package dev.fritz2.components
 
 import dev.fritz2.binding.RootStore
-import dev.fritz2.binding.Scope
 import dev.fritz2.binding.SimpleHandler
 import dev.fritz2.dom.html.Li
 import dev.fritz2.dom.html.RenderContext
+import dev.fritz2.dom.html.Scope
 import dev.fritz2.identification.uniqueId
 import dev.fritz2.styling.*
 import dev.fritz2.styling.params.*
