+++ core/src/jsMain/kotlin/dev/fritz2/dom/tag.kt
@@ -2,6 +2,7 @@ package dev.fritz2.dom
 
 import dev.fritz2.binding.*
 import dev.fritz2.dom.html.RenderContext
+import dev.fritz2.dom.html.Scope
 import dev.fritz2.dom.html.TagContext
 import dev.fritz2.dom.html.render
 import dev.fritz2.lenses.IdProvider
